Campaign Selection v0.1
Working on campaign selection via menu at the top of the screen. I think I need to make a service that will provide whatever options are selected in order to decide what to render.

Work in progress. Not sure if I should bother with the angular-mat stuff or just use bootstrap.